services/synthesis-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
google-cloud-aiplatform